Initial import, from moodle.
The objective of this project is to learn the basics of VFS APIs
(in this case, FUSE).

SVFS is a very simple VFS: whenever you change a file, it keeps
a backup for the next 10 minutes.

The skeleton for the file system is already provided, I think
there'll be mostly some learning but not a lot of code writing.